+++ myems-api/app.py
@@ -8,7 +8,7 @@ from core import energyflowdiagram, privilege, textmessage, distributioncircuit,
     shopfloor, webmessage, distributionsystem, store, emailmessage, tenanttype, wechatmessage, space, gateway, \
     offlinemeter, rule, energycategory, sensor, energyitem, notification, menu, datarepairfile, workingcalendar, \
     microgrid, microgridarchitecturetype, microgridbattery, microgridownertype, \
-    microgridevcharger, microgridgenerator, microgridheatpumps, microgridgrid, command
+    microgridevcharger, microgridgenerator, microgridgrid, microgridheatpump, microgridload,  command
 
 from reports import advancedreport
 from reports import combinedequipmentbatch
@@ -334,9 +334,14 @@ api.add_route('/microgridgrids/{id_}',
               microgridgrid.MicrogridGridItem())
 
 api.add_route('/microgridheatpumps',
-              microgridheatpumps.MicrogridHeatpumpCollection())
+              microgridheatpump.MicrogridHeatpumpCollection())
 api.add_route('/microgridheatpumps/{id_}',
-              microgridheatpumps.MicrogridHeatpumpItem())
+              microgridheatpump.MicrogridHeatpumpItem())
+
+api.add_route('/microgridloads',
+              microgridload.MicrogridLoadCollection())
+api.add_route('/microgridloads/{id_}',
+              microgridload.MicrogridLoadItem())
 
 api.add_route('/microgridownertypes',
               microgridownertype.MicrogridOwnerTypeCollection())